Preserve owned WASM credential mappings
Track shared credential registry mappings by tool owner so WASM
re-registration replaces stale mappings for the same tool without
clobbering another tool that uses the same secret. Keep anonymous
registry additions from losing existing host patterns for that secret.

Move credential injector tests into a sibling test module and split the
shared registry into its own module so production files stay below the
repository line-count guideline. Document that credential persistence
belongs to the registration phase, and make empty CARGO overrides fall
back to the detected cargo binary.
