+++ docs/developers-guide.md
@@ -1179,9 +1179,6 @@ needed, applies overrides, and returns a `PreparedWasmTool`. It can return
 - `apply_wasm_overrides` applies explicit description and schema overrides,
   then attaches runtime-scoped secrets and OAuth configuration. `None` values
   leave the wrapper state unchanged.
-- `persist_credential_mappings` stores HTTP credential mappings after
-  registration succeeds. This is deliberately outside preparation so rejected
-  registrations do not persist credential metadata.
 
 See [ADR 011](adr-011-extract-register-wasm-helpers.md) for the refactoring
 rationale and the maintenance rule behind this helper split.
@@ -1200,6 +1197,11 @@ rationale and the maintenance rule behind this helper split.
    Loads the stored tool record and binary with integrity verification,
    normalizes description and schema via `normalized_schema` and
    `normalized_description`, then delegates to `register_wasm`.
+
+`persist_credential_mappings` belongs to this registration phase, not
+the preparation phase. `register_wasm()` invokes it only after the
+registry accepts the prepared wrapper, and `register_wasm_from_storage()`
+reaches it by delegating through the same lower-level flow.
 
 The storage path is the one that exercises schema normalization, because
 backends may persist placeholder or null schemas that must be stripped
@@ -1315,6 +1317,18 @@ assert_eq!(
 );
 ```
 
+
+## 32. Expected follow-up changes
+
+This guide documents the environment as of the current branch. The
+compile-time reduction plan is still expected to change some of the
+standard commands further, especially around shared extension build
+artifacts and CI duplication.
+
+When those changes land, this guide must be updated in the same branch
+so local setup instructions stay truthful.
+
+## 33. Phased startup pipeline
 
 ## 32. Expected follow-up changes
 
@@ -1784,6 +1798,38 @@ let (addr, _state) = TestGatewayBuilder::new()
 ```
 
 
+## 34. Borrowed newtypes for schema helper arguments
+
+Three lightweight newtype wrappers in `src/tools/tool/schema_helpers.rs` make
+schema and parameter helper signatures explicit without changing the string
+values used in validation error messages.
+
+Caption: Schema helper newtypes.
+
+| Type | Purpose |
+| --- | --- |
+| `ParamName<'a>` | A JSON parameter key expected in tool input |
+| `SchemaPath` | A dot-separated location in a JSON schema |
+| `ToolName<'a>` | A registered tool identifier used as the root strict-schema path |
+
+`ParamName<'a>` and `ToolName<'a>` are zero-cost wrappers over `&'a str`.
+`SchemaPath` owns its path string so nested paths can be constructed while
+descending through schema nodes. All three types implement `From<&str>` and
+`From<&String>` so existing `&str` and `String` call sites continue to compile
+unchanged. `ToolName` additionally converts into `SchemaPath` because the
+strict-schema validator roots its path at the tool name.
+
+Use these types in function signatures that previously accepted a bare `&str`
+or `String` for a parameter name, schema path, or tool name. The types prevent
+accidental argument transposition and make the intent of each parameter clear
+at the call site.
+
+`SchemaPath::child(segment)` returns an owned schema path representing the child
+path `"<parent>.<segment>"`. Use this instead of manual string concatenation
+when descending into nested schema nodes.
+
+These types are re-exported from `src/tools/mod.rs` and are publicly available
+as `crate::tools::{ParamName, SchemaPath, ToolName}`.
 ## 34. Borrowed newtypes for schema helper arguments
 
 Three lightweight newtype wrappers in `src/tools/tool/schema_helpers.rs` make
