+++ src/tools/wasm/credential_injector.rs
@@ -23,7 +23,6 @@
 //! ```
 
 use std::collections::HashMap;
-use std::sync::RwLock;
 
 use crate::secrets::{
     CredentialLocation, CredentialMapping, DecryptedSecret, SecretError, SecretsStore,
@@ -57,139 +56,6 @@ impl From<SecretError> for InjectionError {
             SecretError::DecryptionFailed(msg) => InjectionError::DecryptionFailed(msg),
             _ => InjectionError::DecryptionFailed(e.to_string()),
         }
-    }
-}
-
-/// Thread-safe registry of credential mappings from all installed tools.
-///
-/// Aggregates credential mappings from WASM tools so the built-in HTTP tool can
-/// auto-inject credentials for matching hosts. Uses `std::sync::RwLock` so
-/// `requires_approval` (sync) can query it without async.
-pub struct SharedCredentialRegistry {
-    mappings: RwLock<Vec<CredentialMapping>>,
-}
-
-impl SharedCredentialRegistry {
-    /// Create an empty registry.
-    pub fn new() -> Self {
-        Self {
-            mappings: RwLock::new(Vec::new()),
-        }
-    }
-
-    /// Add or replace credential mappings by secret name.
-    pub fn add_mappings(&self, mappings: impl IntoIterator<Item = CredentialMapping>) {
-        let mappings = dedupe_mappings_by_secret_name(mappings);
-        if mappings.is_empty() {
-            return;
-        }
-
-        match self.mappings.write() {
-            Ok(mut guard) => {
-                replace_mappings_by_secret_name(&mut guard, mappings);
-            }
-            Err(poisoned) => {
-                tracing::warn!(
-                    "SharedCredentialRegistry RwLock poisoned during add_mappings; recovering"
-                );
-                let mut guard = poisoned.into_inner();
-                replace_mappings_by_secret_name(&mut guard, mappings);
-            }
-        }
-    }
-
-    /// Remove all credential mappings whose `secret_name` matches any of the given names.
-    ///
-    /// Called when an extension is unregistered/deactivated so its credential
-    /// injection authority does not outlive the extension.
-    pub fn remove_mappings_for_secrets(&self, secret_names: &[String]) {
-        let mut guard = match self.mappings.write() {
-            Ok(guard) => guard,
-            Err(poisoned) => {
-                tracing::warn!(
-                    "SharedCredentialRegistry RwLock poisoned during remove_mappings_for_secrets; recovering"
-                );
-                poisoned.into_inner()
-            }
-        };
-        guard.retain(|m| !secret_names.contains(&m.secret_name));
-    }
-
-    /// Check if any credential mapping matches this host (sync, for requires_approval).
-    pub fn has_credentials_for_host(&self, host: &str) -> bool {
-        let guard = match self.mappings.read() {
-            Ok(guard) => guard,
-            Err(poisoned) => {
-                tracing::warn!(
-                    "SharedCredentialRegistry RwLock poisoned during has_credentials_for_host; recovering"
-                );
-                poisoned.into_inner()
-            }
-        };
-        guard.iter().any(|mapping| {
-            mapping
-                .host_patterns
-                .iter()
-                .any(|pattern| host_matches_pattern(host, pattern))
-        })
-    }
-
-    /// Get all credential mappings matching a host (for injection).
-    pub fn find_for_host(&self, host: &str) -> Vec<CredentialMapping> {
-        let guard = match self.mappings.read() {
-            Ok(guard) => guard,
-            Err(poisoned) => {
-                tracing::warn!(
-                    "SharedCredentialRegistry RwLock poisoned during find_for_host; recovering"
-                );
-                poisoned.into_inner()
-            }
-        };
-        guard
-            .iter()
-            .filter(|mapping| {
-                mapping
-                    .host_patterns
-                    .iter()
-                    .any(|pattern| host_matches_pattern(host, pattern))
-            })
-            .cloned()
-            .collect()
-    }
-}
-
-fn dedupe_mappings_by_secret_name(
-    mappings: impl IntoIterator<Item = CredentialMapping>,
-) -> Vec<CredentialMapping> {
-    let mappings = mappings.into_iter().collect::<Vec<_>>();
-    let mut deduped = Vec::new();
-    for mapping in mappings.into_iter().rev() {
-        if !deduped
-            .iter()
-            .any(|existing: &CredentialMapping| existing.secret_name == mapping.secret_name)
-        {
-            deduped.push(mapping);
-        }
-    }
-    deduped.reverse();
-    deduped
-}
-
-fn replace_mappings_by_secret_name(
-    guard: &mut Vec<CredentialMapping>,
-    mappings: Vec<CredentialMapping>,
-) {
-    let secret_names = mappings
-        .iter()
-        .map(|mapping| mapping.secret_name.as_str())
-        .collect::<Vec<_>>();
-    guard.retain(|mapping| !secret_names.contains(&mapping.secret_name.as_str()));
-    guard.extend(mappings);
-}
-
-impl Default for SharedCredentialRegistry {
-    fn default() -> Self {
-        Self::new()
     }
 }
 
@@ -364,7 +230,7 @@ pub(crate) fn host_matches_pattern(host: &str, pattern: &str) -> bool {
 }
 
 /// Simple base64 encoding (avoids extra dependency).
-fn base64_encode(input: &[u8]) -> String {
+pub(crate) fn base64_encode(input: &[u8]) -> String {
     const ALPHABET: &[u8] = b"ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789+/";
 
     let mut result = String::new();
@@ -394,292 +260,4 @@ fn base64_encode(input: &[u8]) -> String {
     }
 
     result
-}
-
-#[cfg(test)]
-mod tests {
-    use std::collections::HashMap;
-
-    use crate::secrets::{
-        CreateSecretParams, CredentialLocation, CredentialMapping, InMemorySecretsStore,
-        SecretsStore,
-    };
-    use crate::testing::credentials::{TEST_OPENAI_API_KEY, test_secrets_store};
-    use crate::tools::wasm::credential_injector::{
-        CredentialInjector, base64_encode, host_matches_pattern,
-    };
-
-    fn test_store() -> InMemorySecretsStore {
-        test_secrets_store()
-    }
-
-    #[test]
-    fn test_host_matches_exact() {
-        assert!(host_matches_pattern("api.openai.com", "api.openai.com"));
-        assert!(!host_matches_pattern("api.openai.com", "other.com"));
-    }
-
-    #[test]
-    fn test_host_matches_wildcard() {
-        assert!(host_matches_pattern("api.example.com", "*.example.com"));
-        assert!(host_matches_pattern("sub.api.example.com", "*.example.com"));
-        assert!(!host_matches_pattern("example.com", "*.example.com"));
-    }
-
-    #[test]
-    fn test_base64_encode() {
-        assert_eq!(base64_encode(b"hello"), "aGVsbG8=");
-        assert_eq!(base64_encode(b"user:pass"), "dXNlcjpwYXNz");
-    }
-
-    #[tokio::test]
-    async fn test_inject_bearer() {
-        let store = test_store();
-        store
-            .create(
-                "user1",
-                CreateSecretParams::new("openai_key", TEST_OPENAI_API_KEY),
-            )
-            .await
-            .unwrap();
-
-        let mut mappings = HashMap::new();
-        mappings.insert(
-            "openai".to_string(),
-            CredentialMapping {
-                secret_name: "openai_key".to_string(),
-                location: CredentialLocation::AuthorizationBearer,
-                host_patterns: vec!["api.openai.com".to_string()],
-            },
-        );
-
-        let injector = CredentialInjector::new(mappings, vec!["openai_key".to_string()]);
-        let result = injector
-            .inject("user1", "api.openai.com", &store)
-            .await
-            .unwrap();
-
-        assert_eq!(
-            result.headers.get("Authorization"),
-            Some(&format!("Bearer {TEST_OPENAI_API_KEY}"))
-        );
-    }
-
-    #[tokio::test]
-    async fn test_inject_custom_header() {
-        let store = test_store();
-        store
-            .create("user1", CreateSecretParams::new("api_key", "secret123"))
-            .await
-            .unwrap();
-
-        let mut mappings = HashMap::new();
-        mappings.insert(
-            "custom".to_string(),
-            CredentialMapping {
-                secret_name: "api_key".to_string(),
-                location: CredentialLocation::Header {
-                    name: "X-API-Key".to_string(),
-                    prefix: None,
-                },
-                host_patterns: vec!["*.example.com".to_string()],
-            },
-        );
-
-        let injector = CredentialInjector::new(mappings, vec!["api_key".to_string()]);
-        let result = injector
-            .inject("user1", "api.example.com", &store)
-            .await
-            .unwrap();
-
-        assert_eq!(
-            result.headers.get("X-API-Key"),
-            Some(&"secret123".to_string())
-        );
-    }
-
-    #[tokio::test]
-    async fn test_inject_basic_auth() {
-        let store = test_store();
-        store
-            .create("user1", CreateSecretParams::new("password", "mypassword"))
-            .await
-            .unwrap();
-
-        let mut mappings = HashMap::new();
-        mappings.insert(
-            "basic".to_string(),
-            CredentialMapping {
-                secret_name: "password".to_string(),
-                location: CredentialLocation::AuthorizationBasic {
-                    username: "myuser".to_string(),
-                },
-                host_patterns: vec!["api.service.com".to_string()],
-            },
-        );
-
-        let injector = CredentialInjector::new(mappings, vec!["password".to_string()]);
-        let result = injector
-            .inject("user1", "api.service.com", &store)
-            .await
-            .unwrap();
-
-        // myuser:mypassword base64 encoded
-        let expected = format!("Basic {}", base64_encode(b"myuser:mypassword"));
-        assert_eq!(result.headers.get("Authorization"), Some(&expected));
-    }
-
-    #[tokio::test]
-    async fn test_no_credentials_for_host() {
-        let store = test_store();
-
-        let injector = CredentialInjector::new(HashMap::new(), vec![]);
-        let result = injector
-            .inject("user1", "unknown.com", &store)
-            .await
-            .unwrap();
-
-        assert!(result.is_empty());
-    }
-
-    #[tokio::test]
-    async fn test_access_denied_for_secret() {
-        let store = test_store();
-        store
-            .create("user1", CreateSecretParams::new("secret_key", "value"))
-            .await
-            .unwrap();
-
-        let mut mappings = HashMap::new();
-        mappings.insert(
-            "test".to_string(),
-            CredentialMapping {
-                secret_name: "secret_key".to_string(),
-                location: CredentialLocation::AuthorizationBearer,
-                host_patterns: vec!["api.test.com".to_string()],
-            },
-        );
-
-        // Empty allowed list = nothing allowed
-        let injector = CredentialInjector::new(mappings, vec![]);
-        let result = injector.inject("user1", "api.test.com", &store).await;
-
-        assert!(result.is_err());
-    }
-
-    // ── SharedCredentialRegistry tests ─────────────────────────────────
-
-    use crate::tools::wasm::credential_injector::SharedCredentialRegistry;
-
-    #[test]
-    fn test_shared_registry_empty() {
-        let registry = SharedCredentialRegistry::new();
-        assert!(!registry.has_credentials_for_host("api.example.com"));
-        assert!(registry.find_for_host("api.example.com").is_empty());
-    }
-
-    #[test]
-    fn test_shared_registry_add_and_find() {
-        let registry = SharedCredentialRegistry::new();
-        registry.add_mappings(vec![
-            CredentialMapping::bearer("openai_key", "api.openai.com"),
-            CredentialMapping::header("github_token", "X-GitHub-Token", "*.github.com"),
-        ]);
-
-        assert!(registry.has_credentials_for_host("api.openai.com"));
-        assert!(!registry.has_credentials_for_host("api.anthropic.com"));
-
-        let found = registry.find_for_host("api.openai.com");
-        assert_eq!(found.len(), 1);
-        assert_eq!(found[0].secret_name, "openai_key");
-    }
-
-    #[test]
-    fn test_shared_registry_wildcard_host() {
-        let registry = SharedCredentialRegistry::new();
-        registry.add_mappings(vec![CredentialMapping::bearer("gh_token", "*.github.com")]);
-
-        assert!(registry.has_credentials_for_host("api.github.com"));
-        assert!(registry.has_credentials_for_host("uploads.github.com"));
-        assert!(!registry.has_credentials_for_host("github.com"));
-    }
-
-    #[test]
-    fn test_shared_registry_multiple_adds() {
-        let registry = SharedCredentialRegistry::new();
-        registry.add_mappings(vec![CredentialMapping::bearer("key1", "api.example.com")]);
-        registry.add_mappings(vec![CredentialMapping::bearer("key2", "api.example.com")]);
-
-        let found = registry.find_for_host("api.example.com");
-        assert_eq!(found.len(), 2);
-    }
-
-    #[test]
-    fn test_shared_registry_replaces_mappings_with_same_secret_name() {
-        let registry = SharedCredentialRegistry::new();
-        registry.add_mappings(vec![CredentialMapping::bearer("key1", "old.example.com")]);
-        registry.add_mappings(vec![CredentialMapping::bearer("key1", "api.example.com")]);
-
-        assert!(registry.find_for_host("old.example.com").is_empty());
-        let found = registry.find_for_host("api.example.com");
-        assert_eq!(found.len(), 1);
-        assert_eq!(found[0].secret_name, "key1");
-    }
-
-    #[test]
-    fn test_shared_registry_remove_mappings_for_secrets() {
-        let registry = SharedCredentialRegistry::new();
-        registry.add_mappings(vec![
-            CredentialMapping::bearer("openai_key", "api.openai.com"),
-            CredentialMapping::bearer("gh_token", "*.github.com"),
-            CredentialMapping::header("openai_org", "OpenAI-Organization", "api.openai.com"),
-        ]);
-
-        assert_eq!(registry.find_for_host("api.openai.com").len(), 2);
-        assert!(registry.has_credentials_for_host("api.github.com"));
-
-        // Remove only mappings for openai secrets
-        registry.remove_mappings_for_secrets(&["openai_key".to_string(), "openai_org".to_string()]);
-
-        // OpenAI mappings should be gone
-        assert!(registry.find_for_host("api.openai.com").is_empty());
-        // GitHub mapping should remain
-        assert!(registry.has_credentials_for_host("api.github.com"));
-    }
-
-    #[test]
-    fn test_shared_registry_remove_nonexistent_is_noop() {
-        let registry = SharedCredentialRegistry::new();
-        registry.add_mappings(vec![CredentialMapping::bearer("key1", "api.example.com")]);
-
-        registry.remove_mappings_for_secrets(&["nonexistent".to_string()]);
-        assert_eq!(registry.find_for_host("api.example.com").len(), 1);
-    }
-
-    #[test]
-    fn test_shared_registry_thread_safety() {
-        use std::sync::Arc;
-        use std::thread;
-
-        let registry = Arc::new(SharedCredentialRegistry::new());
-
-        let handles: Vec<_> = (0..4)
-            .map(|i| {
-                let r = Arc::clone(&registry);
-                thread::spawn(move || {
-                    r.add_mappings(vec![CredentialMapping::bearer(
-                        format!("key_{}", i),
-                        "api.example.com",
-                    )]);
-                })
-            })
-            .collect();
-
-        for h in handles {
-            h.join().unwrap();
-        }
-
-        let found = registry.find_for_host("api.example.com");
-        assert_eq!(found.len(), 4);
-    }
 }
